pyproject.toml
[project]
name = "vlm2vec-benchmark"
version = "0.1.0"
description = "Colab-ready evaluation harness for multi-vector retrievers on MMEB-V2"
authors = [{ name = "Logan Mann" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch~=2.2",
  "transformers~=4.44",
  "accelerate~=0.33",
  "pillow~=10.3",
  "torchvision~=0.17",
  "tqdm~=4.66",
  "numpy~=1.26",
  "pyyaml~=6.0",
  "datasets~=2.20",
  "huggingface-hub~=0.23",
  "safetensors~=0.4",
  "sentencepiece~=0.2",
  "einops~=0.8",
  "timm~=1.0",
  "pandas~=2.2",
  "rich~=13.7",
  "keybert~=0.7",
  "umap-learn~=0.5",
  "hdbscan~=0.8",
  "bertopic~=0.16",
  "evaluate~=0.4",
  "pynvml~=11.5",
  "pytest~=8.3",
  "pytest-cov~=4.1"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["src", "."]
